BAD NEWS
Reached a point in this where I wanted to redesign the dashboard using a datatable instead and a refresher. However the data from crypto compare api is absolutely dog shit. The data isn't good enough to feel good about providing information to public. The data supplied is so under-updated. So slow, so bad. Just with a simple A -> A arbitrage check, the data is old, exchanges are inactive/old/bad. Data is bad. Not going to work. Coin market cap is currently $375 /mo for the data required. Would require writing a custom wrapper to integrate to it.